Super cool script for connecting the Sony WH-1000 headset
- On my Oracle Linux machine, connecting to the Sony wh1000 headset via the
settings UI is annoying, sometimes it enters the low quality headset mode,
sometimes it plainly doesn't connect.
- Also, there appears to be two volume controls for the WH, one is remembered on
the headset, Windows, Android and Fedora machines that I have set this volume when
moving the volume sliders, Oracle Linux doesn't. So I have to tap it out on the
physical headset control, lame-o...
- Instead of fixing the issue properly, let's apply some duct tape.
- First check whether the MAC provided as the argument is connected via
hcitool con output.
- If connected, get the proper DBUS object for volume settings, sometimes this
object appears a few seconds after the connection pops up, so retry the operation.
- Send the volume 100, this is not the max, seems like it's 127, however, I read
it's possible to set it too high and get some distortion, yuck, don't wanna deal
with that. 100 is fine, it's fine!
- Use pulse audio to detect the output sink for our mac address and then set this
sink to be the default one.